Changing sphinx_log id default to bigint(20)
This https://github.com/kaltura/server/blob/dbd28230f0298fb2a94120bfd2990835c27c3907/plugins/search/data/sql/schema.sql#L14 should have also changed here, but didn't

+++ deployment/base/sql/01.kaltura_sphinx_ce_tables.sql
@@ -6,7 +6,7 @@ USE `kaltura_sphinx_log`;
 SET GLOBAL sql_mode = '';
 /*Table structure for table `sphinx_log` */
 CREATE TABLE IF NOT EXISTS `sphinx_log` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
+  `id` bigint(20) NOT NULL AUTO_INCREMENT,
   `executed_server_id` int(11) NOT NULL,
   `object_type` varchar(255) NOT NULL,
   `object_id` varchar(20) NOT NULL,
